05-competitive-intel/system/requirements.txt
anthropic>=0.40.0
firecrawl-py>=1.0.0
pyyaml>=6.0
weasyprint>=62.0
